Refactored statistic functions to their own files and updated docstrings
